Create LSTM/CNN network that can process multiple samples for one target

Creates a network consisting of an arbitrary number of CNN, LSTM and dense layers with multiple input layers. After LSTM/CNN part all representations get aggregated by summation. Can be used to make single prediction for combination of multiple input sequences. Implements approach as described here

create_model_lstm_cnn_multi_input(
  maxlen = 50,
  dropout_lstm = 0,
  recurrent_dropout_lstm = 0,
  layer_lstm = NULL,
  layer_dense = c(4),
  solver = "adam",
  learning_rate = 0.001,
  vocabulary_size = 4,
  bidirectional = FALSE,
  batch_size = NULL,
  compile = TRUE,
  kernel_size = NULL,
  filters = NULL,
  strides = NULL,
  pool_size = NULL,
  padding = "same",
  dilation_rate = NULL,
  gap = FALSE,
  use_bias = TRUE,
  zero_mask = FALSE,
  label_smoothing = 0,
  label_noise_matrix = NULL,
  last_layer_activation = "softmax",
  loss_fn = "categorical_crossentropy",
  num_input_layers = 1,
  auc_metric = FALSE,
  f1_metric = FALSE,
  samples_per_target,
  batch_norm_momentum = 0.99,
  verbose = TRUE,
  model_seed = NULL
)

Arguments

maxlen

Length of predictor sequence.

dropout_lstm

Fraction of the units to drop for inputs.

recurrent_dropout_lstm

Fraction of the units to drop for recurrent state.

layer_lstm

Number of cells per network layer. Can be a scalar or vector.

layer_dense

Vector specifying number of neurons per dense layer after last LSTM or CNN layer (if no LSTM used).

solver

Optimization method, options are "adam", "adagrad", "rmsprop" or "sgd".

learning_rate

Learning rate for optimizer.

vocabulary_size

Number of unique character in vocabulary.

bidirectional

Use bidirectional wrapper for lstm layers.

batch_size

Number of samples that are used for one network update. Only used if stateful = TRUE.

compile

Whether to compile the model.

kernel_size

Size of 1d convolutional layers. For multiple layers, assign a vector. (e.g, rep(3,2) for two layers and kernel size 3)

filters

Number of filters. For multiple layers, assign a vector.

strides

Stride values. For multiple layers, assign a vector.

pool_size

Integer, size of the max pooling windows. For multiple layers, assign a vector.

padding

Padding of CNN layers, e.g. "same", "valid" or "causal".

dilation_rate

Integer, the dilation rate to use for dilated convolution.

gap

Whether to apply global average pooling after last CNN layer.

use_bias

Boolean. Usage of bias for CNN layers.

zero_mask

Boolean, whether to apply zero masking before LSTM layer. Only used if model does not use any CNN layers.

label_smoothing

Float in [0, 1]. If 0, no smoothing is applied. If > 0, loss between the predicted labels and a smoothed version of the true labels, where the smoothing squeezes the labels towards 0.5. The closer the argument is to 1 the more the labels get smoothed.

label_noise_matrix

Matrix of label noises. Every row stands for one class and columns for percentage of labels in that class. If first label contains 5 percent wrong labels and second label no noise, then

label_noise_matrix <- matrix(c(0.95, 0.05, 0, 1), nrow = 2, byrow = TRUE )

last_layer_activation

Either "sigmoid" or "softmax".

loss_fn

Either "categorical_crossentropy" or "binary_crossentropy". If label_noise_matrix given, will use custom "noisy_loss".

num_input_layers

Number of input layers.

auc_metric

Whether to add AUC metric.

f1_metric

Whether to add F1 metric.

samples_per_target

Number of samples to combine for one target.

batch_norm_momentum

Momentum for the moving mean and the moving variance.

verbose

Boolean.

model_seed

Set seed for model parameters in tensorflow if not NULL.

Examples

create_model_lstm_cnn_multi_input(
  maxlen = 50,
  vocabulary_size = 4,
  samples_per_target = 7,
  kernel_size = c(10, 10),
  filters = c(64, 128),
  pool_size = c(2, 2),
  layer_lstm = c(32),
  layer_dense = c(64, 2),
  learning_rate = 0.001)
#> Model: "model_3"
#> ________________________________________________________________________________
#>  Layer (type)         Output Shape   Param #  Connected to           Trainable  
#> ================================================================================
#>  input_1 (InputLayer)  [(None, 50, 4  0       []                     Y          
#>                       )]                                                        
#>  input_2 (InputLayer)  [(None, 50, 4  0       []                     Y          
#>                       )]                                                        
#>  input_3 (InputLayer)  [(None, 50, 4  0       []                     Y          
#>                       )]                                                        
#>  input_4 (InputLayer)  [(None, 50, 4  0       []                     Y          
#>                       )]                                                        
#>  input_5 (InputLayer)  [(None, 50, 4  0       []                     Y          
#>                       )]                                                        
#>  input_6 (InputLayer)  [(None, 50, 4  0       []                     Y          
#>                       )]                                                        
#>  input_7 (InputLayer)  [(None, 50, 4  0       []                     Y          
#>                       )]                                                        
#>  model_2 (Functional)  (None, 32)    106048   ['input_1[0][0]',      Y          
#>                                                'input_2[0][0]',                 
#>                                                'input_3[0][0]',                 
#>                                                'input_4[0][0]',                 
#>                                                'input_5[0][0]',                 
#>                                                'input_6[0][0]',                 
#>                                                'input_7[0][0]']                 
#>  add (Add)            (None, 32)     0        ['model_2[0][0]',      Y          
#>                                                'model_2[1][0]',                 
#>                                                'model_2[2][0]',                 
#>                                                'model_2[3][0]',                 
#>                                                'model_2[4][0]',                 
#>                                                'model_2[5][0]',                 
#>                                                'model_2[6][0]']                 
#>  dense_4 (Dense)      (None, 64)     2112     ['add[0][0]']          Y          
#>  dense_5 (Dense)      (None, 2)      130      ['dense_4[0][0]']      Y          
#> ================================================================================
#> Total params: 108,290
#> Trainable params: 107,906
#> Non-trainable params: 384
#> ________________________________________________________________________________