Collect samples from generator and store in rds or pickle file.

Repeatedly generate samples with data generator and store output. Creates a separate rds or pickle file in output_path for each batch.

dataset_from_gen(
  output_path,
  iterations = 10,
  train_type = "lm",
  output_format = "target_right",
  path_corpus,
  batch_size = 32,
  maxlen = 250,
  step = NULL,
  vocabulary = c("a", "c", "g", "t"),
  as_numpy_array = FALSE,
  shuffle = FALSE,
  set_learning = NULL,
  seed = NULL,
  random_sampling,
  file_name_start = "batch_",
  ...
)

Arguments

output_path

Output directory. Output files will be named output_path + file_name_start + x + ".rds" or ".pickle", where x is an index (from 1 to iterations) and file ending depends on as_numpy_array argument.

output_format

Determines shape of output tensor for language model. Either "target_right", "target_middle_lstm", "target_middle_cnn" or "wavenet". Assume a sequence "AACCGTA". Output correspond as follows

"target_right": X = "AACCGT", Y = "A"

"target_middle_lstm": X = (X_1 = "AAC", X_2 = "ATG"), Y = "C" (note reversed order of X_2)

"target_middle_cnn": X = "AACGTA", Y = "C"

"wavenet": X = "AACCGT", Y = "ACCGTA"

path_corpus

Input directory where .fasta files are located or path to single file ending with .fasta or .fastq (as specified in format argument). Can also be a list of directories and/or files.

batch_size

Number of samples in one batch.

maxlen

Length of predictor sequence.

step

How often to take a sample.

vocabulary

Vector of allowed characters, character outside vocabulary get encoded as 0-vector.

as_numpy_array

Store output as list of numpy arrays if TRUE (otherwise as R array).

shuffle

Whether to shuffle samples within each batch.

set_learning

When you want to assign one label to set of samples. Only implemented for train_type = "label_folder". Input is a list with the following parameters

samples_per_target: how many samples to use for one target

maxlen: length of one sample

reshape_mode: "time_dist", "multi_input" or "concat".

If reshape_mode is "multi_input", generator will produce samples_per_target separate inputs, each of length maxlen (model should have samples_per_target input layers).

If reshape_mode is "time_dist", generator will produce a 4D input array. The dimensions correspond to (batch_size, samples_per_target, maxlen, length(vocabulary)).

If reshape mode is "concat", generator will concatenate samples_per_target sequences of length maxlen to one long sequence

If reshape_mode is "concat", there is an additional buffer_len argument. If buffer_len is an integer, the subsequences are interspaced with buffer_len rows. The input length is (maxlen \(*\) samples_per_target) + buffer_len \(*\) (samples_per_target - 1)

seed

Sets seed for `set.seed`` function for reproducible results.

file_name_start

Start of output file names.

...

further generator options.

Examples

# create dummy fasta files
temp_dir <- tempfile()
dir.create(temp_dir)
create_dummy_data(file_path = temp_dir,
                  num_files = 3,
                  seq_length = 8, 
                  num_seq = 2)

# extract samples
out_dir <- tempfile()
dir.create(out_dir)
dataset_from_gen(output_path = out_dir,
                 iterations = 10,
                 train_type = "lm",
                 output_format = "target_right",
                 path_corpus = temp_dir, 
                 batch_size = 32,
                 maxlen = 5,
                 step = 1,
                 file_name_start = "batch_")

list.files(out_dir)
#>  [1] "batch_1.rds"  "batch_10.rds" "batch_2.rds"  "batch_3.rds"  "batch_4.rds" 
#>  [6] "batch_5.rds"  "batch_6.rds"  "batch_7.rds"  "batch_8.rds"  "batch_9.rds" 

# read first batch
z <- readRDS(list.files(out_dir, full.names = TRUE)[1])
x <- z[[1]]
y <- z[[2]]
dim(x)
#> [1] 32  5  4
dim(y)
#> [1] 32  4