Model hyperparameter callback

Get model hyperparameters.

hyper_param_with_model_cb(
  default_arguments,
  model,
  path_tensorboard,
  run_name,
  train_type,
  path_model,
  path,
  train_val_ratio,
  batch_size,
  epochs,
  max_queue_size,
  lr_plateau_factor,
  patience,
  cooldown,
  steps_per_epoch,
  step,
  shuffle_file_order,
  initial_epoch,
  vocabulary,
  learning_rate,
  shuffle_input,
  vocabulary_label,
  solver,
  file_limit,
  reverse_complement,
  wavenet_format,
  cnn_format
)

Arguments

model

A keras model.

path_tensorboard

Path to tensorboard directory or NULL. If NULL, training not tracked on tensorboard.

run_name

Name of the run. Name will be used to identify output from callbacks. If NULL, will use date as run name. If name already present, will add "_2" to name or "_{x+1}" if name ends with _x, where x is some integer.

train_type

Either "lm", "lm_rds" for language model; "label_header", "label_folder", "label_csv", "label_rds" for classification or "dummy_gen".

Language model is trained to predict character(s) in a sequence.

"label_header"/"label_folder"/"label_csv" are trained to predict a corresponding class given a sequence as input.

If "label_header", class will be read from fasta headers.

If "label_folder", class will be read from folder, i.e. all files in one folder must belong to the same class.

If "label_csv", targets are read from a csv file. This file should have one column named "file". The targets then correspond to entries in that row (except "file" column). Example: if we are currently working with a file called "a.fasta" and coresponding label is "label_1", there should be a row in our csv file

file	label_1	label_2

"a.fasta"	1	0

If "label_rds", generator will iterate over set of .rds files containing each a list of input and target tensors. Not implemented for model with multiple inputs.

If "lm_rds", generator will iterate over set of .rds files and will split tensor according to target_len argument (targets are last target_len nucleotides of each sequence).

If "dummy_gen", generator creates random data once and repeatedly feeds these to model.

path

Path to training data. If train_type is label_folder, should be a vector or list where each entry corresponds to a class (list elements can be directories and/or individual files). If train_type is not label_folder, can be a single directory or file or a list of directories and/or files.

train_val_ratio

For generator defines the fraction of batches that will be used for validation (compared to size of training data), i.e. one validation iteration processes batch_size \(*\) steps_per_epoch \(*\) train_val_ratio samples. If you use dataset instead of generator and dataset_val is NULL, splits dataset into train/validation data.

batch_size

Number of samples used for one network update.

epochs

Number of iterations.

max_queue_size

Maximum size for the generator queue.

lr_plateau_factor

Factor of decreasing learning rate when plateau is reached.

patience

Number of epochs waiting for decrease in validation loss before reducing learning rate.

cooldown

Number of epochs without changing learning rate.

steps_per_epoch

Number of training batches per epoch.

step

Frequency of sampling steps.

shuffle_file_order

Boolean, whether to go through files sequentially or shuffle beforehand.

initial_epoch

Epoch at which to start training. Note that network will run for (epochs - initial_epochs) rounds and not epochs rounds.

vocabulary

Vector of allowed characters. Characters outside vocabulary get encoded as specified in ambiguous_nuc.

shuffle_input

Whether to shuffle entries in file.

vocabulary_label

Character vector of possible targets. Targets outside vocabulary_label will get discarded if train_type = "label_header".

file_limit

Integer or NULL. If integer, use only specified number of randomly sampled files for training. Ignored if greater than number of files in path.

reverse_complement

Boolean, for every new file decide randomly to use original data or its reverse complement.