Encodes integer sequence for language model

Helper function for generator_fasta_lm. Encodes integer sequence to input/target list according to output_format argument.

seq_encoding_lm(
  sequence = NULL,
  maxlen,
  vocabulary,
  start_ind,
  ambiguous_nuc = "zero",
  nuc_dist = NULL,
  use_quality = FALSE,
  quality_vector = NULL,
  target_len = 1,
  use_coverage = FALSE,
  max_cov = NULL,
  cov_vector = NULL,
  n_gram = NULL,
  n_gram_stride = 1,
  output_format = "target_right",
  char_sequence = NULL,
  adjust_start_ind = FALSE,
  tokenizer = NULL
)

Arguments

sequence

Sequence of integers.

maxlen

Length of one sample

vocabulary

Vector of allowed characters, character outside vocabulary get encoded as 0-vector.

start_ind

Start positions of samples in sequence.

ambiguous_nuc

How to handle nucleotides outside vocabulary, either "zero", "empirical" or "equal". See train_model. Note that "discard" option is not available for this function.

nuc_dist

Nucleotide distribution.

use_quality

Use quality scores.

quality_vector

Vector of quality probabilities.

target_len

Number of nucleotides to predict at once.

use_coverage

Whether to use coverage encoding.

max_cov

Biggest coverage value. Only applies if use_coverage = TRUE.

cov_vector

Vector of coverage values associated to the input.

n_gram

Integer, encode target not nucleotide wise but combine n nucleotides at once. For example for n=2, "AA" -> (1, 0,..., 0), "AC" -> (0, 1, 0,..., 0), "TT" -> (0,..., 0, 1), where the one-hot vectors have length length(vocabulary)^n.

n_gram_stride

Step size for n-gram encoding. For AACCGGTT with n-gram = 4 and n_gram_stride = 2, generator encodes (AACC), (CCGG), (GGTT); for n_gram_stride = 4 generator encodes (AACC), (GGTT).

output_format

Determines shape of output tensor for language model. Either "target_right", "target_middle_lstm", "target_middle_cnn" or "wavenet". Assume a sequence "AACCGTA". Output correspond as follows

"target_right": X = "AACCGT", Y = "A"

"target_middle_lstm": X = (X_1 = "AAC", X_2 = "ATG"), Y = "C" (note reversed order of X_2)

"target_middle_cnn": X = "AACGTA", Y = "C"

"wavenet": X = "AACCGT", Y = "ACCGTA"

char_sequence

A character string.

adjust_start_ind

Whether to shift values in start_ind to start at 1: for example (5,11,25) becomes (1,7,21).

tokenizer

A keras tokenizer.

Examples

# use integer sequence as input 

z <- seq_encoding_lm(sequence = c(1,0,5,1,3,4,3,1,4,1,2),
maxlen = 5,
vocabulary = c("a", "c", "g", "t"),
start_ind = c(1,3),
ambiguous_nuc = "equal",
target_len = 1,
output_format = "target_right")

x <- z[[1]]
y <- z[[2]]

x[1,,] # 1,0,5,1,3
#>      [,1] [,2] [,3] [,4]
#> [1,] 1.00 0.00 0.00 0.00
#> [2,] 0.00 0.00 0.00 0.00
#> [3,] 0.25 0.25 0.25 0.25
#> [4,] 1.00 0.00 0.00 0.00
#> [5,] 0.00 0.00 1.00 0.00
y[1,] # 4
#> [1] 0 0 0 1

x[2,,] # 5,1,3,4,
#>      [,1] [,2] [,3] [,4]
#> [1,] 0.25 0.25 0.25 0.25
#> [2,] 1.00 0.00 0.00 0.00
#> [3,] 0.00 0.00 1.00 0.00
#> [4,] 0.00 0.00 0.00 1.00
#> [5,] 0.00 0.00 1.00 0.00
y[2,] # 1
#> [1] 1 0 0 0

# use character string as input
z <- seq_encoding_lm(sequence = NULL,
maxlen = 5,
vocabulary = c("a", "c", "g", "t"),
start_ind = c(1,3),
ambiguous_nuc = "zero",
target_len = 1,
output_format = "target_right",
char_sequence = "ACTaaTNTNaZ")

x <- z[[1]]
y <- z[[2]]

x[1,,] # actaa
#>      [,1] [,2] [,3] [,4]
#> [1,]    1    0    0    0
#> [2,]    0    1    0    0
#> [3,]    0    0    0    1
#> [4,]    1    0    0    0
#> [5,]    1    0    0    0
y[1,] # t
#> [1] 0 0 0 1

x[2,,] # taatn
#>      [,1] [,2] [,3] [,4]
#> [1,]    0    0    0    1
#> [2,]    1    0    0    0
#> [3,]    1    0    0    0
#> [4,]    0    0    0    1
#> [5,]    0    0    0    0
y[2,] # t
#> [1] 0 0 0 1