Create wavenet model

Creates network architecture as described here. Implementation uses code from here.

create_model_wavenet(
  filters = 16,
  kernel_size = 2,
  residual_blocks,
  maxlen,
  input_tensor = NULL,
  initial_kernel_size = 32,
  initial_filters = 32,
  output_channels = 4,
  output_activation = "softmax",
  solver = "adam",
  learning_rate = 0.001,
  compile = TRUE,
  verbose = TRUE,
  model_seed = NULL
)

Arguments

filters

Integer, the dimensionality of the output space (i.e. the number of output filters in the convolution).

kernel_size

An integer or list of a single integer, specifying the length of the 1D convolution window.

residual_blocks

if a single integer: the number of residual blocks in the model - the dilation rate of each block `i` is calculated by `2^i`. If a vector, then it's used as the dilation rate of each residual block.

maxlen

Length of predictor sequence.

input_tensor

Tensor to use as input for the model. Usually a 3d tensor.

initial_kernel_size

kernel size of the first causal convolution.

initial_filters

number of filters in the initial causal convolution.

output_channels

number of channels in the output.

output_activation

activation function for the last layer. Default: 'softmax'

solver

Optimization method, options are "adam", "adagrad", "rmsprop" or "sgd".

learning_rate

Learning rate for optimizer.

compile

Whether to compile the model.

verbose

Boolean.

model_seed

Set seed for model parameters in tensorflow if not NULL.

Examples

model <- create_model_wavenet(residual_blocks = 2^rep(1:4, 2), maxlen = 1000)
#> Model: "model_7"
#> ________________________________________________________________________________
#>  Layer (type)             Output Shape      Param #  Connected to               
#> ================================================================================
#>  input_8 (InputLayer)     [(None, 1000, 4)  0        []                         
#>                           ]                                                     
#>  conv1d_13 (Conv1D)       (None, 1000, 32)  4096     ['input_8[0][0]']          
#>  r_layer (RLayer)         [(None, 1000, 32  0        ['conv1d_13[0][0]']        
#>                           ),                                                    
#>                            (None, 1000, 32                                      
#>                           )]                                                    
#>  r_layer_1 (RLayer)       [(None, 1000, 32  0        ['r_layer[0][0]']          
#>                           ),                                                    
#>                            (None, 1000, 32                                      
#>                           )]                                                    
#>  r_layer_2 (RLayer)       [(None, 1000, 32  0        ['r_layer_1[0][0]']        
#>                           ),                                                    
#>                            (None, 1000, 32                                      
#>                           )]                                                    
#>  r_layer_3 (RLayer)       [(None, 1000, 32  0        ['r_layer_2[0][0]']        
#>                           ),                                                    
#>                            (None, 1000, 32                                      
#>                           )]                                                    
#>  r_layer_4 (RLayer)       [(None, 1000, 32  0        ['r_layer_3[0][0]']        
#>                           ),                                                    
#>                            (None, 1000, 32                                      
#>                           )]                                                    
#>  r_layer_5 (RLayer)       [(None, 1000, 32  0        ['r_layer_4[0][0]']        
#>                           ),                                                    
#>                            (None, 1000, 32                                      
#>                           )]                                                    
#>  r_layer_6 (RLayer)       [(None, 1000, 32  0        ['r_layer_5[0][0]']        
#>                           ),                                                    
#>                            (None, 1000, 32                                      
#>                           )]                                                    
#>  r_layer_7 (RLayer)       [(None, 1000, 32  0        ['r_layer_6[0][0]']        
#>                           ),                                                    
#>                            (None, 1000, 32                                      
#>                           )]                                                    
#>  add_1 (Add)              (None, 1000, 32)  0        ['r_layer[0][1]',          
#>                                                       'r_layer_1[0][1]',        
#>                                                       'r_layer_2[0][1]',        
#>                                                       'r_layer_3[0][1]',        
#>                                                       'r_layer_4[0][1]',        
#>                                                       'r_layer_5[0][1]',        
#>                                                       'r_layer_6[0][1]',        
#>                                                       'r_layer_7[0][1]']        
#>  activation (Activation)  (None, 1000, 32)  0        ['add_1[0][0]']            
#>  conv1d_15 (Conv1D)       (None, 1000, 16)  512      ['activation[0][0]']       
#>  conv1d_14 (Conv1D)       (None, 1000, 4)   68       ['conv1d_15[0][0]']        
#> ================================================================================
#> Total params: 4,676
#> Trainable params: 4,676
#> Non-trainable params: 0
#> ________________________________________________________________________________