Read states from h5 file

Reads h5 file created by predict_model function.

load_prediction(
  h5_path,
  rows = NULL,
  verbose = FALSE,
  get_sample_position = FALSE,
  get_seq = FALSE
)

Arguments

h5_path

Path to h5 file.

rows

Range of rows to read. If NULL read everything.

verbose

Boolean.

get_sample_position

Return position of sample corresponding to state if TRUE.

get_seq

Return nucleotide sequence if TRUE.

Examples

# make prediction for single sequence and write to h5 file
model <- create_model_lstm_cnn(maxlen = 20, layer_lstm = 8, layer_dense = 2, verbose = FALSE)
vocabulary <- c("a", "c", "g", "t")
sequence <- paste(sample(vocabulary, 200, replace = TRUE), collapse = "")
output_file <- tempfile(fileext = ".h5")
predict_model(output_format = "one_seq", model = model, step = 10,
              sequence = sequence, filename = output_file, mode = "label")
#> layer_name not specified. Using layer dense_19
#> Computing output for model at layer dense_19 
#> Model: "model_15"
#> ________________________________________________________________________________
#>  Layer (type)                       Output Shape                    Param #     
#> ================================================================================
#>  input_15 (InputLayer)              [(None, 20, 4)]                 0           
#>  lstm_15 (LSTM)                     (None, 8)                       416         
#>  dense_19 (Dense)                   (None, 2)                       18          
#> ================================================================================
#> Total params: 434
#> Trainable params: 434
#> Non-trainable params: 0
#> ________________________________________________________________________________
#> NULL
load_prediction(h5_path = output_file)
#> $states
#>            [,1]      [,2]
#>  [1,] 0.4635138 0.5364861
#>  [2,] 0.4408394 0.5591606
#>  [3,] 0.4601055 0.5398945
#>  [4,] 0.4372636 0.5627365
#>  [5,] 0.4982560 0.5017441
#>  [6,] 0.4225605 0.5774395
#>  [7,] 0.4331294 0.5668706
#>  [8,] 0.4903475 0.5096526
#>  [9,] 0.4773623 0.5226377
#> [10,] 0.5130779 0.4869222
#> [11,] 0.4370126 0.5629873
#> [12,] 0.4708784 0.5291216
#> [13,] 0.4764718 0.5235282
#> [14,] 0.4526482 0.5473517
#> [15,] 0.4872468 0.5127532
#> [16,] 0.4784909 0.5215091
#> [17,] 0.5007567 0.4992432
#> [18,] 0.4271716 0.5728284
#> [19,] 0.4436159 0.5563841
#>