Make prediction for nucleotide sequence or entries in fasta/fastq file

Removes layers (optional) from pretrained model and calculates states of fasta/fastq file or nucleotide sequence. Writes states to h5 or csv file (access content of h5 output with load_predictions function). There are several options on how to process an input file:

If "one_seq", computes prediction for sequence argument or fasta/fastq file. Combines fasta entries in file to one sequence. This means predictor sequences can contain elements from more than one fasta entry.

If "by_entry", will output a separate file for each fasta/fastq entry. Names of output files are: output_dir + "Nr" + i + filename + output_type, where i is the number of the fasta entry.

If "by_entry_one_file", will store prediction for all fasta entries in one h5 file.

If "one_pred_per_entry", will make one prediction for each entry by either picking random sample for long sequences or pad sequence for short sequences.

predict_model(
  output_format = "one_seq",
  model = NULL,
  layer_name = NULL,
  sequence = NULL,
  path_input = NULL,
  round_digits = 2,
  filename = "states.h5",
  step = 1,
  vocabulary = c("a", "c", "g", "t"),
  batch_size = 256,
  verbose = TRUE,
  return_states = FALSE,
  padding_maxlen = FALSE,
  output_type = "h5",
  path_model = NULL,
  mode = "lm",
  lm_format = "target_right",
  output_dir = NULL,
  format = "fasta",
  include_seq = TRUE,
  reverse_complement_encoding = FALSE
)

Arguments

output_format

Either "one_seq", "by_entry", "by_entry_one_file", "one_pred_per_entry".

model

A keras model. If model and path_model are not NULL, model will be used for inference.

layer_name

Name of layer to get output from. If NULL, will use the last layer.

sequence

Character string, ignores path_input if argument given.

path_input

Path to fasta file.

round_digits

Number of decimal places.

filename

Filename to store states in. No file output if argument is NULL.

step

Frequency of sampling steps.

vocabulary

Vector of allowed characters, character outside vocabulary get encoded as 0-vector.

batch_size

Number of samples to evaluate at once. Does not change output, only relevant for speed and memory.

verbose

Boolean.

return_states

Return predictions as data frame. Only supported for output_format "one_seq".

padding_maxlen

Boolean, whether to pad input with maxlen 0-vectors. (First prediction is for tensor of zeros)

output_type

"h5" or "csv". If output_format`` is "by_entries_one_file", "one_pred_per_entry"can only be"h5"`.

path_model

Path to a pretrained model.

mode

Either "lm" for language model or "label" for label classification.

output_dir

Path to folder, where to write output.

format

Either "fasta" or "fastq".

include_seq

Whether to include input sequence in h5 file.

Examples

# make prediction for single sequence and write to h5 file
model <- create_model_lstm_cnn(maxlen = 20, layer_lstm = 8, layer_dense = 2, verbose = FALSE)
vocabulary <- c("a", "c", "g", "t")
sequence <- paste(sample(vocabulary, 200, replace = TRUE), collapse = "")
output_file <- tempfile(fileext = ".h5")
predict_model(output_format = "one_seq", model = model, step = 10,
             sequence = sequence, filename = output_file, mode = "label")
#> layer_name not specified. Using layer dense_25
#> Computing output for model at layer dense_25 
#> Model: "model_21"
#> ________________________________________________________________________________
#>  Layer (type)                       Output Shape                    Param #     
#> ================================================================================
#>  input_17 (InputLayer)              [(None, 20, 4)]                 0           
#>  lstm_19 (LSTM)                     (None, 8)                       416         
#>  dense_25 (Dense)                   (None, 2)                       18          
#> ================================================================================
#> Total params: 434
#> Trainable params: 434
#> Non-trainable params: 0
#> ________________________________________________________________________________
#> NULL

# make prediction for fasta file with multiple entries, write output to separate h5 files
fasta_path <- tempfile(fileext = ".fasta")
create_dummy_data(file_path = fasta_path, num_files = 1,
                 num_seq = 5, seq_length = 100,
                 write_to_file_path = TRUE)
model <- create_model_lstm_cnn(maxlen = 20, layer_lstm = 8, layer_dense = 2, verbose = FALSE)
output_dir <- tempfile()
dir.create(output_dir)
predict_model(output_format = "by_entry", model = model, step = 10, verbose = FALSE,
               output_dir = output_dir, mode = "label", path_input = fasta_path)
list.files(output_dir)
#> [1] "states_nr_1.h5" "states_nr_2.h5" "states_nr_3.h5" "states_nr_4.h5"
#> [5] "states_nr_5.h5"